smart-talk-agent/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
langchain>=1.0.0
langchain-openai>=0.3.0
langgraph>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
sentence-transformers>=3.0.0
numpy>=1.26.0
pyyaml>=6.0
python-multipart>=0.0.9
